fix: fix circular dependency lookup in RpcControllerFactory
While we were correctly detecting a circular dependency condition within
the factory method itself, we needed to prevent a circular _lookup_ as
well, which occurs when the plugin manager calls `has()` on the service.

This patch adds a check to see if the service being checked is the same
as the last one being created; if so, it returns true to ensure we do
not attempt a second creation.